+++ utils/common/src/main/java/org/apache/brooklyn/util/http/HttpAsserts.java
@@ -156,7 +156,7 @@ public class HttpAsserts {
     public static void assertHttpStatusCodeEquals(String url, int... acceptableReturnCodes) {
         List<Integer> acceptableCodes = Lists.newArrayList();
         for (int code : acceptableReturnCodes) {
-            acceptableCodes.add((Integer)code);
+            acceptableCodes.add(code);
         }
         try {
             int actualCode = HttpTool.getHttpStatusCode(url);
